+++ CHANGELOG.md
@@ -6,6 +6,10 @@
 
 * [#1591](https://github.com/bbatsov/projectile/issues/1591): Add `project.el` integration that will make Projectile the default provider for project lookup.
 
+### Bug fixed
+
+* [#1799](https://github.com/bbatsov/projectile/pull/1799): Fix `projectile-open-projects` lists projects for which all buffers are closed.
+
 ## 2.6.0 (2022-10-25)
 
 ### New features